[*] changed pointsToPixels rounding

+++ auxiliary/UiScaling.h
@@ -9,7 +9,7 @@ class UiScaling
 
 public:
 
-    static int pointsToPixels(float a_pointsValue);
+    static float pointsToPixels(float a_pointsValue);
     static QSize pointsToPixels(const QSize& a_pointsSize);
 
     inline static double pointsToInches(float a_pointsValue);
